+++ change-log.md
@@ -7,6 +7,12 @@
 
 ## Change log
 
+### v2026.4.107
+Released on 03/29/2026
+* BUGFIX: Fix sync loss after suspend/resume by reloading the webview once network is restored (https://github.com/patrikx3/onenote/issues/204).
+* FEATURE: Toast now supports sticky mode (stays until clicked or replaced by a new toast).
+* FEATURE: Added safe publish workflow with automatic electron dependency restore on interruption.
+
 ### v2026.4.105
 Released on 03/28/2026
 * FEATURE: Refactored the Electron main process to ESM (.mjs) with import/export syntax.
@@ -82,7 +88,7 @@ All my domains, including [patrikx3.com](https://patrikx3.com), [corifeus.eu](ht
 ---
 
 
-[**P3X-ONENOTE**](https://corifeus.com/onenote) Build v2026.4.106
+[**P3X-ONENOTE**](https://corifeus.com/onenote) Build v2026.4.107
 
  [![NPM](https://img.shields.io/npm/v/p3x-onenote.svg)](https://www.npmjs.com/package/p3x-onenote)  [![Donate for PatrikX3 / P3X](https://img.shields.io/badge/Donate-PatrikX3-003087.svg)](https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QZVM4V6HVZJW6)  [![Contact Corifeus / P3X](https://img.shields.io/badge/Contact-P3X-ff9900.svg)](https://www.patrikx3.com/en/front/contact) [![Like Corifeus @ Facebook](https://img.shields.io/badge/LIKE-Corifeus-3b5998.svg)](https://www.facebook.com/corifeus.software)